feat: implementa EquipamentoService

+++ src/main/java/com/gestaooserp/dev/repository/EquipamentoRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.Equipamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipamentoRepository extends JpaRepository<Equipamento,Integer> {
+public interface EquipamentoRepository extends JpaRepository<Equipamento,Long> {
 }
